backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
pyyaml==6.0.1
aiofiles==23.2.1
cors==1.0.1
Pillow==10.0.0
bcrypt==4.1.1
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.8.1
